refactor: 테스트 클래스 도메인 패키지 안으로 이동
